requirements.txt
streamlit
pdfplumber
python-docx
wordcloud
matplotlib
groq